+++ lib/ooxml_parser/common_parser/common_data/alternate_content/drawing/graphic/shape/shape_properties/custom_geometry/docx_custom_geometry/docx_shape_line_path/docx_shape_line_element.rb
@@ -23,8 +23,6 @@ module OoxmlParser
         @type = :arc
       when 'cubicBezTo'
         @type = :cubic_bezier
-      when 'quadBezTo'
-        @type = :quadratic_bezier
       when 'close'
         @type = :close
       end