+++ lib/ooxml_parser/common_parser/common_data/alternate_content/picture/shape/old_docx_shape_properties.rb
@@ -3,7 +3,7 @@
 module OoxmlParser
   # Fallback DOCX Shape Properties
   class OldDocxShapeProperties < OOXMLDocumentObject
-    attr_accessor :fill_color, :opacity, :stroke_color, :stroke_weight
+    attr_accessor :fill_color, :stroke_color, :stroke_weight
 
     # Parse OldDocxShapeProperties object
     # @param node [Nokogiri::XML:Element] node to parse
@@ -13,8 +13,6 @@ module OoxmlParser
         case key
         when 'fillcolor'
           @fill_color = Color.new(parent: self).parse_hex_string(value.value.delete('#'))
-        when 'opacity'
-          @opacity = value.value.to_f
         when 'strokecolor'
           @stroke_color = Color.new(parent: self).parse_hex_string(value.value.delete('#'))
         when 'strokeweight'