+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/docx_paragraph/docx_paragraph_run.rb
@@ -121,8 +121,6 @@ module OoxmlParser
               @shape = Shape.new(parent: self).parse(pict_node_child, :shape)
             when 'rect'
               @shape = Shape.new(parent: self).parse(pict_node_child, :rectangle)
-            when 'oval'
-              @shape = Shape.new(parent: self).parse(pict_node_child, :oval)
             end
           end
         when 'object'