+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/header_footer.rb
@@ -32,8 +32,6 @@ module OoxmlParser
       when 'endnoteReference'
         @path_suffix = 'endnote'
         @type = :footer
-      else
-        warn "Unknown HeaderFooter type: #{node.name}"
       end
     end