+++ spec/common/classes/ooxml_document_object_spec.rb
@@ -8,4 +8,12 @@ describe OoxmlParser::OOXMLDocumentObject do
       expect(OoxmlParser::BookmarkStart.new).not_to eq(OoxmlParser::BookmarkEnd.new)
     end
   end
+
+  describe 'Class methods' do
+    describe 'OOXMLDocumentObject.encrypted_file?' do
+      it 'OOXMLDocumentObject.encrypted_file?(path, ignore_system: true) always return false' do
+        expect(described_class).not_to be_encrypted_file(__FILE__, ignore_system: true)
+      end
+    end
+  end
 end